Rename: components/recoil 폴더 삭제
- cart.ts state폴더로 이동
