+++ components/layouts/CartItemList.tsx
@@ -2,7 +2,7 @@ import React, { Dispatch, SetStateAction, useEffect, useState } from "react";
 import CartItem from "../ui/CartItem";
 import { DefaultValue, useRecoilState } from "recoil";
 import { cartListType } from "@/types/types";
-import { frozenCartListState, generalCartListState } from "../recoil/cart";
+import { frozenCartListState, generalCartListState } from "../../state/cart";
 import Link from "next/link";
 import Image from "next/image";
