+++ state/cart.ts
@@ -1,7 +1,7 @@
 import { FROZEN_CART_LIST, GENERAL_CART_LIST } from "@/data/StaticData";
 import { cartListType } from "@/types/types";
 import { atom } from "recoil";
-import CartItem from "../ui/CartItem";
+import CartItem from "../components/ui/CartItem";
 import { recoilPersist } from "recoil-persist";
 
 const { persistAtom } = recoilPersist({
